chore: adjust bestmove scaling (#311)
Adjust scaling of best move stability in the case that the best move changes during a search. We now provide a small time bonus (10%) when best move stability goes to 0.

STC regression:

Elo   | -0.95 +- 1.84 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.96 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 64220 W: 19091 L: 19267 D: 25862
Penta | [1924, 7553, 13260, 7521, 1852]
https://openbench.nocturn9x.space/test/6553/

LTC regression:

Elo   | 0.53 +- 2.97 (95%)
SPRT  | 40.0+0.40s Threads=1 Hash=128MB
LLR   | 2.99 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 21732 W: 5984 L: 5951 D: 9797
Penta | [506, 2485, 4818, 2584, 473]
https://openbench.nocturn9x.space/test/6557/

bench: 845875

+++ chess/src/move_generation/metadata.rs
@@ -119,9 +119,9 @@ pub fn compute(board: &Board) -> CheckPinMetadata {
             let mut checkers_clone = checkers;
             let checker = bitboard_helpers::next_bit(&mut checkers_clone) as u8;
 
-            let ray = rays::between(king_sq, checker as u8);
+            let ray = rays::between(king_sq, checker);
 
-            if let Some((piece, side)) = board.piece_on_square(checker as u8) {
+            if let Some((piece, side)) = board.piece_on_square(checker) {
                 debug_assert!(side == them);
                 let is_slider = piece.is_slider();
                 if is_slider {
